Add files via upload
In the "app-debug.apk" you can see the Widget that updates each 30 minutes or when you press the update button.
The connection to the webserver is stored as a shared preference for the widget and can be modified any time.

The screenshots ilustrate the basic of this proyect. With more effort more aesthetic is achievable.